refactor event to share pipelinevent with stageevent
