+++ src/main/java/io/jenkins/plugins/StageEvent.java
@@ -3,13 +3,20 @@ package io.jenkins.plugins;
 import java.util.Objects;
 
 public class StageEvent {
+  private PipelineEvent pipelineEvent;
   private String stageName;
 
-  public StageEvent() {
+  public PipelineEvent getPipelineEvent() {
+    return this.pipelineEvent;
   }
 
-  public StageEvent(String stageName) {
-    this.stageName = stageName;
+  public void setPipelineEvent(PipelineEvent pipelineEvent) {
+    this.pipelineEvent = pipelineEvent;
+  }
+
+  public StageEvent pipelineEvent(PipelineEvent pipelineEvent) {
+    this.pipelineEvent = pipelineEvent;
+    return this;
   }
 
   public String getStageName() {
